Module Protobuf.Any

`Any` contains an arbitrary serialized protocol buffer message along with a URL that describes the type of the serialized message.

Protobuf library provides support to pack/unpack Any values in the form of utility functions or additional generated methods of the Any type.

Example 1: Pack and unpack a message in C++.

           Foo foo = ...;
           Any any;
           any.PackFrom(foo);
           ...
           if (any.UnpackTo(&foo)) {
             ...
           }

Example 2: Pack and unpack a message in Java.

           Foo foo = ...;
           Any any = Any.pack(foo);
           ...
           if (any.is(Foo.class)) {
             foo = any.unpack(Foo.class);
           }

Example 3: Pack and unpack a message in Python.

           foo = Foo(...)
           any = Any()
           any.Pack(foo)
           ...
           if any.Is(Foo.DESCRIPTOR):
             any.Unpack(foo)
             ...

Example 4: Pack and unpack a message in Go

            foo := &pb.Foo{...}
            any, err := anypb.New(foo)
            if err != nil {
              ...
            }
            ...
            foo := &pb.Foo{}
            if err := any.UnmarshalTo(foo); err != nil {
              ...
            }

The pack methods provided by protobuf library will by default use 'type.googleapis.com/full.type.name' as the type URL and the unpack methods only use the fully qualified type name after the last '/' in the type URL, for example "foo.bar.com/x/y.z" will yield type name "y.z".

JSON

The JSON representation of an `Any` value uses the regular representation of the deserialized, embedded message, with an additional field `@type` which contains the type URL. Example:

           package google.profile;
           message Person {
             string first_name = 1;
             string last_name = 2;
           }

           {
             "@type": "type.googleapis.com/google.profile.Person",
             "firstName": <string>,
             "lastName": <string>
           }

If the embedded message type is well-known and has a custom JSON representation, that representation will be embedded adding a field `value` which holds the custom JSON in addition to the `@type` field. Example (for message [google.protobuf.Duration][]):

           {
             "@type": "type.googleapis.com/google.protobuf.Duration",
             "value": "1.212s"
           }

type t = {

type_url : string;

value : bytes;
(*
Must be a valid serialized protocol buffer of the above specified type.
*)

}

val make : ?type_url:string -> ?value:bytes -> unit -> t

Helper function to generate a message using default values

val to_proto : t -> Ocaml_protoc_plugin.Writer.t

Serialize the message to binary format

val from_proto : Ocaml_protoc_plugin.Reader.t -> (t, [> Ocaml_protoc_plugin.Result.error ]) Stdlib.result

Deserialize from binary format

val to_json : Ocaml_protoc_plugin.Json_options.t -> t -> Ocaml_protoc_plugin.Json.t

Serialize to Json (compatible with Yojson.Basic.t)

val from_json : Ocaml_protoc_plugin.Json.t -> (t, [> Ocaml_protoc_plugin.Result.error ]) Stdlib.result

Deserialize from Json (compatible with Yojson.Basic.t)

val name : unit -> string

Fully qualified protobuf name of this message