Module FieldDescriptorProto.Type

type t =

| TYPE_DOUBLE
(*
0 is reserved for errors. Order is weird for historical reasons.
*)

| TYPE_FLOAT

| TYPE_INT64
(*
Not ZigZag encoded. Negative numbers take 10 bytes. Use TYPE_SINT64 if negative values are likely.
*)

| TYPE_UINT64

| TYPE_INT32
(*
Not ZigZag encoded. Negative numbers take 10 bytes. Use TYPE_SINT32 if negative values are likely.
*)

| TYPE_FIXED64

| TYPE_FIXED32

| TYPE_BOOL

| TYPE_STRING

| TYPE_GROUP
(*
Tag-delimited aggregate. Group type is deprecated and not supported in proto3. However, Proto3 implementations should still be able to parse the group wire format and treat group fields as unknown fields.
*)

| TYPE_MESSAGE
(*
Length-delimited aggregate.
*)

| TYPE_BYTES
(*
New in version 2.
*)

| TYPE_UINT32

| TYPE_ENUM

| TYPE_SFIXED32

| TYPE_SFIXED64

| TYPE_SINT32
(*
Uses ZigZag encoding.
*)

| TYPE_SINT64
(*
Uses ZigZag encoding.
*)

val name : unit -> string

Fully qualified protobuf name of this enum