Module Protobuf.FileOptions

module OptimizeMode : sig ... end

Generated classes can be optimized for speed or code size.

type t = {

java_package : string option;

java_outer_classname : string option;

optimize_for : OptimizeMode.t;

java_multiple_files : bool;

go_package : string option;

cc_generic_services : bool;

java_generic_services : bool;

py_generic_services : bool;

java_generate_equals_and_hash : bool option;
(*

alert protobuf Marked as deprecated in the .proto file

*)

deprecated : bool;
(*
Is this file deprecated? Depending on the target platform, this can emit Deprecated annotations for everything in the file, or it will be completely ignored; in the very least, this is a formalization for deprecating files.
*)

java_string_check_utf8 : bool;

cc_enable_arenas : bool;

objc_class_prefix : string option;

csharp_namespace : string option;

swift_prefix : string option;

php_class_prefix : string option;

php_namespace : string option;

php_generic_services : bool;

php_metadata_namespace : string option;

ruby_package : string option;

uninterpreted_option : UninterpretedOption.t list;

extensions' : Ocaml_protoc_plugin.Extensions.t;

}

val make : ?java_package:string -> ?java_outer_classname:string -> ?optimize_for:OptimizeMode.t -> ?java_multiple_files:bool -> ?go_package:string -> ?cc_generic_services:bool -> ?java_generic_services:bool -> ?py_generic_services:bool -> ?java_generate_equals_and_hash:bool -> ?deprecated:bool -> ?java_string_check_utf8:bool -> ?cc_enable_arenas:bool -> ?objc_class_prefix:string -> ?csharp_namespace:string -> ?swift_prefix:string -> ?php_class_prefix:string -> ?php_namespace:string -> ?php_generic_services:bool -> ?php_metadata_namespace:string -> ?ruby_package:string -> ?uninterpreted_option:UninterpretedOption.t list -> ?extensions':Ocaml_protoc_plugin.Extensions.t -> unit -> t

Helper function to generate a message using default values

val to_proto : t -> Ocaml_protoc_plugin.Writer.t

Serialize the message to binary format

val from_proto : Ocaml_protoc_plugin.Reader.t -> (t, [> Ocaml_protoc_plugin.Result.error ]) Stdlib.result

Deserialize from binary format

val to_json : Ocaml_protoc_plugin.Json_options.t -> t -> Ocaml_protoc_plugin.Json.t

Serialize to Json (compatible with Yojson.Basic.t)

val from_json : Ocaml_protoc_plugin.Json.t -> (t, [> Ocaml_protoc_plugin.Result.error ]) Stdlib.result

Deserialize from Json (compatible with Yojson.Basic.t)

val name : unit -> string

Fully qualified protobuf name of this message