Module GeneratedCodeInfo.Annotation

type t = {

path : int list;
(*
Identifies the element in the original source .proto file. This field is formatted the same as SourceCodeInfo.Location.path.
*)

source_file : string option;

begin' : int option;
(*
Identifies the starting offset in bytes in the generated code that relates to the identified object.
*)

end' : int option;
(*
Identifies the ending offset in bytes in the generated code that relates to the identified offset. The end offset should be one past the last relevant byte (so the length of the text = end - begin).
*)

}

val make : ?path:int list -> ?source_file:string -> ?begin':int -> ?end':int -> unit -> t

Helper function to generate a message using default values

val to_proto : t -> Ocaml_protoc_plugin.Writer.t

Serialize the message to binary format

val from_proto : Ocaml_protoc_plugin.Reader.t -> (t, [> Ocaml_protoc_plugin.Result.error ]) Stdlib.result

Deserialize from binary format

val to_json : Ocaml_protoc_plugin.Json_options.t -> t -> Ocaml_protoc_plugin.Json.t

Serialize to Json (compatible with Yojson.Basic.t)

val from_json : Ocaml_protoc_plugin.Json.t -> (t, [> Ocaml_protoc_plugin.Result.error ]) Stdlib.result

Deserialize from Json (compatible with Yojson.Basic.t)

val name : unit -> string

Fully qualified protobuf name of this message