Module SourceCodeInfo.Location

type t = {

path : int list;
(*
Identifies which part of the FileDescriptorProto was defined at this location.

Each element is a field number or an index. They form a path from the root FileDescriptorProto to the place where the definition occurs. For example, this path:

               [ 4, 3, 2, 7, 1 ]

refers to:

               file.message_type(3)  // 4, 3
                   .field(7)         // 2, 7
                   .name()           // 1

This is because FileDescriptorProto.message_type has field number 4:

               repeated DescriptorProto message_type = 4;

and DescriptorProto.field has field number 2:

               repeated FieldDescriptorProto field = 2;

and FieldDescriptorProto.name has field number 1:

               optional string name = 1;

Thus, the above path gives the location of a field name. If we removed the last element:

               [ 4, 3, 2, 7 ]

this path refers to the whole field declaration (from the beginning of the label to the terminating semicolon).
*)

span : int list;
(*
Always has exactly three or four elements: start line, start column, end line (optional, otherwise assumed same as start line), end column. These are packed into a single field for efficiency. Note that line and column numbers are zero-based -- typically you will want to add 1 to each before displaying to a user.
*)

leading_comments : string option;

trailing_comments : string option;

leading_detached_comments : string list;

}

val make : ?path:int list -> ?span:int list -> ?leading_comments:string -> ?trailing_comments:string -> ?leading_detached_comments:string list -> unit -> t

Helper function to generate a message using default values

val to_proto : t -> Ocaml_protoc_plugin.Writer.t

Serialize the message to binary format

val from_proto : Ocaml_protoc_plugin.Reader.t -> (t, [> Ocaml_protoc_plugin.Result.error ]) Stdlib.result

Deserialize from binary format

val to_json : Ocaml_protoc_plugin.Json_options.t -> t -> Ocaml_protoc_plugin.Json.t

Serialize to Json (compatible with Yojson.Basic.t)

val from_json : Ocaml_protoc_plugin.Json.t -> (t, [> Ocaml_protoc_plugin.Result.error ]) Stdlib.result

Deserialize from Json (compatible with Yojson.Basic.t)

val name : unit -> string

Fully qualified protobuf name of this message