Module Google.Protobuf

module NullValue : sig ... end

`NullValue` is a singleton enumeration to represent the null value for the `Value` type union.

module Struct : sig ... end

`Struct` represents a structured data value, consisting of fields which map to dynamically typed values. In some languages, `Struct` might be supported by a native representation. For example, in scripting languages like JS a struct is represented as an object. The details of that representation are described together with the proto support for the language.

module Value : sig ... end

`Value` represents a dynamically typed value which can be either null, a number, a string, a boolean, a recursive struct value, or a list of values. A producer of value is expected to set one of these variants. Absence of any variant indicates an error.

module ListValue : sig ... end

`ListValue` is a wrapper around a repeated field of values.