Module Protobuf

module Spec : sig ... end

module Serialize : sig ... end

module Deserialize : sig ... end

module Reader : sig ... end

Some buffer to hold data, and to read and write data

module Writer : sig ... end

Some buffer to hold data, and to read and write data

module Service : sig ... end