Module Protobuf.Reader

Some buffer to hold data, and to read and write data

type t = {

mutable offset : Base.int;

end_offset : Base.int;

data : Base.String.t;

}

type error = [

| `Premature_end_of_input

| `Unknown_field_type of Base.int

]

val pp_error : Ppx_deriving_runtime.Format.formatter -> error -> Ppx_deriving_runtime.unit

val show_error : error -> Ppx_deriving_runtime.string

val create : ?⁠offset:Base.int -> ?⁠length:Base__Int.t -> Base.String.t -> t

val size : t -> Base__Int.t

val validate_capacity : t -> Base__Int.t -> (unit, [> `Premature_end_of_input ]) Base.Result.t

Return an error if there is not enough data in input

val has_more : t -> bool

Test if there is more data in the buffer to be read

val read_byte : t -> (int, [> `Premature_end_of_input ]) Base__Result.t

val read_raw_varint : t -> (int, [> `Premature_end_of_input ]) Base__Result.t

val read_varint : t -> (Spec.field, [> `Premature_end_of_input ]) Base__Result.t

val read_field_header : t -> (Base.int * Base.int, error) Base.Result.t

val read_length_delimited : t -> (Spec.field, [> `Premature_end_of_input ]) Base__Result.t

val read_fixed32 : t -> (Spec.field, [> `Premature_end_of_input ]) Base__Result.t

val read_fixed64 : t -> (Spec.field, [> `Premature_end_of_input ]) Base__Result.t

val read_field : t -> (Base.int * Spec.field, error) Base.Result.t