Module Protobuf__Deserialize

type error = [

| Protobuf.Reader.error

| `Wrong_field_type of Base.string * Protobuf.Spec.field

| `Illegal_value of Base.string * Protobuf.Spec.field

| `Not_implemented

| `Unknown_enum_value of Base.int

| `Oneof_missing

]

val pp_error : Ppx_deriving_runtime.Format.formatter -> error -> Ppx_deriving_runtime.unit

val show_error : error -> Ppx_deriving_runtime.string

type nonrec 'a result = ('a, error) Base.Result.t

Module for deserializing values

type _ spec =

| Double : Base.float spec

| Float : Base.float spec

| Int32 : Base.int spec

| Int64 : Base.int spec

| UInt32 : Base.int spec

| UInt64 : Base.int spec

| SInt32 : Base.int spec

| SInt64 : Base.int spec

| Fixed32 : Base.int spec

| Fixed64 : Base.int spec

| SFixed32 : Base.int spec

| SFixed64 : Base.int spec

| Bool : Base.bool spec

| String : Base.string spec

| Bytes : Base.bytes spec

| Enum : (Base.int -> 'a result) -> 'a spec

type _ compound =

| Message : (Protobuf.Reader.t -> 'a result) -> 'a Base.option compound

| RepeatedMessage : (Protobuf.Reader.t -> 'a result) -> 'a Base.list compound

| Repeated : 'a spec -> 'a Base.list compound

| Basic : 'a spec -> 'a compound

type 'a sentinal = Base.unit -> 'a

type decoder = Protobuf.Spec.field -> Base.unit result

val error_wrong_field : Base.string -> Protobuf.Spec.field -> 'a result

val error_illegal_value : Base.string -> Protobuf.Spec.field -> 'a result

val read_fields : Protobuf.Reader.t -> (Base.int * Protobuf.Spec.field) Base.list result

Deserialize a reader.

val deserialize : (Base.int * decoder) Base.list -> Protobuf.Reader.t -> Base.unit result

Deserialize takes a list of field sentinals and a reader for the data. Data is decoded into a list of tags, fields and the appropiate function for the tag is called. After all fields have been processed, data can be read from the sentinals, created by the caller

A helper function exists to create sentinals

val read_varint : signed:bool -> type_name:Base.string -> Protobuf.Spec.field -> (Base__Int.t, error) Base__Result.t

val ok_exn : ('a, 'b) Base.result -> 'a

val type_of_spec : a. 'a spec -> 'a * [> `Fixed_32_bit | `Fixed_64_bit | `Length_delimited | `Varint ] * (Protobuf.Spec.field -> 'a result)

val sentinal : a. 'a compound -> 'a sentinal * decoder

type _ oneof =

| Oneof : (Base.int * 'b compound * ('b -> 'a)) -> 'a oneof

Oneofs are handled a bit special.

val oneof_sentinal : 'a oneof Base.list -> (Base.unit -> 'a result) * (Base.int * decoder) Base.list