ocaml-protoc-plugin index

Library ocaml-protoc-plugin

The entry point of this library is the module: Ocaml_protoc_plugin.

Library ocaml-protoc-plugin.google_types

The entry point of this library is the module: Google_types.